Resolved Changes(2)
