+++ assets/images.js
@@ -1,0 +1,48 @@
+const images = [
+    {
+        key: 1,
+        src: "https://images.unsplash.com/photo-1433838552652-f9a46b332c40?ixlib=rb-4.0.3&ixid=MnwxMjA3fDB8MHxleHBsb3JlLWZlZWR8NHx8fGVufDB8fHx8&auto=format&fit=crop&w=600&q=60",
+        alt: "Cosmos Flower",
+        width: 320,
+        height: 213,
+        caption: "Cosmos Flower Macro (Tom Eversley - isorepublic.com)",
+    },
+    {
+        key: 2,
+        src: "https://images.unsplash.com/photo-1454391304352-2bf4678b1a7a?ixlib=rb-4.0.3&ixid=MnwxMjA3fDB8MHxleHBsb3JlLWZlZWR8MXx8fGVufDB8fHx8&auto=format&fit=crop&w=600&q=60",
+        width: 271,
+        height: 320,
+        caption: "Orange Macro (Tom Eversley - isorepublic.com)",
+    },
+    {
+        key: 3,
+        src: "https://c1.staticflickr.com/9/8330/28941240416_71d2a7af8e_b.jpg",
+        width: 320,
+        height: 213,
+        tags: [
+            { value: "Nature", title: "Nature" },
+            { value: "People", title: "People" },
+        ],
+        caption: "Surfer Sunset (Tom Eversley - isorepublic.com)",
+    },
+    {
+        key: 4,
+        src: "https://images.unsplash.com/photo-1545389336-cf090694435e?ixlib=rb-4.0.3&ixid=MnwxMjA3fDB8MHxleHBsb3JlLWZlZWR8M3x8fGVufDB8fHx8&auto=format&fit=crop&w=600&q=60",
+        width: 320,
+        height: 213,
+        tags: [
+            { value: "People", title: "People" },
+            { value: "Sport", title: "Sport" },
+        ],
+        caption: "Man on BMX (Tom Eversley - isorepublic.com)",
+    },
+    {
+        key: 5,
+        src: "https://images.unsplash.com/photo-1568323993144-20d546ba585d?ixlib=rb-4.0.3&ixid=MnwxMjA3fDB8MHxleHBsb3JlLWZlZWR8Mnx8fGVufDB8fHx8&auto=format&fit=crop&w=600&q=60",
+        width: 320,
+        height: 213,
+        caption: "Ropeman - Thailand (Tom Eversley - isorepublic.com)",
+    },
+];
+
+export default images;
